+++ app/helpers/application_helper.rb
@@ -158,12 +158,10 @@ module ApplicationHelper
   def link_to_message(message, options={}, html_options = nil)
     link_to(
       h(truncate(message.subject, :length => 60)),
-      { :controller => '/messages', :action => 'show',
-        :board_id => message.board_id,
-        :id => message.root,
-        :r => (message.parent_id && message.id),
-        :anchor => (message.parent_id ? "message-#{message.id}" : nil)
-      }.merge(options),
+      topic_path(message.root,
+                 { :r => (message.parent_id && message.id),
+                   :anchor => (message.parent_id ? "message-#{message.id}" : nil)
+                 }.merge(options)),
       html_options
     )
   end
@@ -995,7 +993,12 @@ module ApplicationHelper
       I18n.locale = "#{I18n.locale}";
     })
     unless User.current.pref.warn_on_leaving_unsaved == '0'
-      tags += javascript_tag("jQuery(function(){ new WarnLeavingUnsaved('#{escape_javascript( l(:text_warn_on_leaving_unsaved) )}'); });")
+      tags += javascript_tag("jQuery(document).ready(function(){
+                                new WarnLeavingUnsaved('#{escape_javascript( l(:text_warn_on_leaving_unsaved) )}');
+                                jQuery(document).ajaxComplete(function(){
+                                    new WarnLeavingUnsaved('#{escape_javascript( l(:text_warn_on_leaving_unsaved) )}')
+                                });
+                            })")
     end
 
     if User.current.impaired? and accessibility_js_enabled?